requirements.txt
reflex
google-genai
pandas
python-dotenv
anthropic

reflex==0.8.25